refactor: move func to contact module
